+++ server.js
@@ -1,2 +1,0 @@
-// TODO: Move to `bin` and use commander
-require('./server/app')(8000);