+++ server/routes/writeJsonToImage.js
@@ -42,12 +42,6 @@ module.exports = function writeTextToConnections (req, res) {
           conn.res.write(buff);
         });
       }
-      // If we have firstConnections, write a header for them
-      if (firstConnections.length) {
-        gif.on('data', writeToFirstConnections);
-        gif.writeHeader();
-        gif.flushData();
-      }
 
       // Process the image (addFrame#1)
       console.log('RAW-ANALYZE: Analyzing image');
@@ -56,6 +50,7 @@ module.exports = function writeTextToConnections (req, res) {
 
       // Write out the image info for the first connections (addFrame#2)
       console.log('RAW-FIRST-INFO: Writing image info to first connections');
+      gif.on('data', writeToFirstConnections);
       gif.writeImageInfo();
       gif.flushData();
       console.log('RAW-FIRST-INFO: First conections completed');